[PyTorch] Split StaticModule out of test_static_runtime
I want to use StaticModule in another (internal) test, so splitting it out.

Differential Revision: [D54384817](https://our.internmc.facebook.com/intern/diff/D54384817/)

[ghstack-poisoned]
